+++ lib/src/presentation/pages/main_page/main_page.dart
@@ -3,11 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:kkw_blog/resource/l10n/generated/l10n.dart';
 
-import 'local_widgets/introduction.dart';
+import 'local_widgets/desktop_view/layout.dart' as Desktop;
 import 'local_widgets/theme_mode_fab.dart';
-import 'local_widgets/blog_title.dart';
-import 'local_widgets/classified_panel.dart';
-import 'local_widgets/preview_listview.dart';
 
 class MainPage extends HookWidget {
   const MainPage({super.key});
@@ -28,22 +25,7 @@ class MainPage extends HookWidget {
           child: Scrollbar(
             thumbVisibility: true,
             controller: scrollController,
-            child: Center(
-              child: ConstrainedBox(
-                constraints: const BoxConstraints(maxWidth: 1200),
-                child: CustomScrollView(
-                  controller: scrollController,
-                  scrollBehavior: ScrollConfiguration.of(context)
-                      .copyWith(scrollbars: false),
-                  slivers: const [
-                    _HorizontalPaddingSliver(),
-                    SliverToBoxAdapter(child: _Header()),
-                    _BodyAreaSliver(),
-                    _HorizontalPaddingSliver(),
-                  ],
-                ),
-              ),
-            ),
+            child: Desktop.Layout(scrollController: scrollController),
           ),
         ),
         floatingActionButton: const ThemeModeFab(),
@@ -64,57 +46,5 @@ class MainPage extends HookWidget {
     if (0 < newPose && newPose < controller.position.maxScrollExtent) {
       controller.jumpTo(controller.offset + event.scrollDelta.dy);
     }
-  }
-}
-
-class _Header extends StatelessWidget {
-  const _Header();
-
-  @override
-  Widget build(BuildContext context) {
-    return const Column(
-      children: [
-        BlogTitle(),
-        SizedBox(height: 10),
-        Introduction(),
-      ],
-    );
-  }
-}
-
-class _BodyAreaSliver extends StatelessWidget {
-  const _BodyAreaSliver();
-
-  @override
-  Widget build(BuildContext context) {
-    return const SliverCrossAxisGroup(
-      slivers: [
-        SliverCrossAxisExpanded(
-          flex: 1,
-          sliver: SliverToBoxAdapter(
-            child: ClassifiedPanel(),
-          ),
-        ),
-        SliverCrossAxisExpanded(
-          flex: 3,
-          sliver: PreviewListview(),
-        ),
-        SliverCrossAxisExpanded(
-          flex: 1,
-          sliver: SliverToBoxAdapter(),
-        ),
-      ],
-    );
-  }
-}
-
-class _HorizontalPaddingSliver extends StatelessWidget {
-  const _HorizontalPaddingSliver();
-
-  @override
-  Widget build(BuildContext context) {
-    return const SliverToBoxAdapter(
-      child: SizedBox(height: kToolbarHeight),
-    );
   }
 }
